requirements.txt
torch
numpy<2.0
pandas
python-dateutil
yfinance
tabulate
wandb
